extracted package object
