+++ dynamodb-json/src/main/scala/zio/dynamodb/json/DynamodbJsonCodec.scala
@@ -4,7 +4,6 @@ import zio.Chunk
 import zio.dynamodb._
 import zio.json._
 import zio.json.ast.Json
-import zio.schema.Schema
 
 import scala.util.Try
 
@@ -110,24 +109,24 @@ object DynamodbJsonCodec {
       }
   }
 
-  implicit class AttrMapJsonOps(am: AttrMap) {
-    def toJsonString: String = Encoder.attributeValueToJsonString(am.toAttributeValue)
-  }
-
-  implicit class ProductJsonOps[A](a: A)(implicit schema: Schema[A]) {
-    def toJsonString: String = Encoder.attributeValueToJsonString(toItem(a).toAttributeValue)
-  }
-
-  private[dynamodb] def toItem[A](a: A)(implicit schema: Schema[A]): Item =
-    FromAttributeValue.attrMapFromAttributeValue
-      .fromAttributeValue(AttributeValue.encode(a)(schema))
-      .getOrElse(throw new Exception(s"error encoding $a"))
-
-  def parse(json: String): Either[DynamoDBError.ItemError, AttrMap] =
-    DynamodbJsonCodec.Decoder
-      .jsonStringToAttributeValue(json)
-      .left
-      .map(DynamoDBError.ItemError.DecodingError)
-      .flatMap(FromAttributeValue.attrMapFromAttributeValue.fromAttributeValue)
+//  implicit class AttrMapJsonOps(am: AttrMap) {
+//    def toJsonString: String = Encoder.attributeValueToJsonString(am.toAttributeValue)
+//  }
+//
+//  implicit class ProductJsonOps[A](a: A)(implicit schema: Schema[A]) {
+//    def toJsonString: String = Encoder.attributeValueToJsonString(toItem(a).toAttributeValue)
+//  }
+//
+//  private[dynamodb] def toItem[A](a: A)(implicit schema: Schema[A]): Item =
+//    FromAttributeValue.attrMapFromAttributeValue
+//      .fromAttributeValue(AttributeValue.encode(a)(schema))
+//      .getOrElse(throw new Exception(s"error encoding $a"))
+//
+//  def parse(json: String): Either[DynamoDBError.ItemError, AttrMap] =
+//    DynamodbJsonCodec.Decoder
+//      .jsonStringToAttributeValue(json)
+//      .left
+//      .map(DynamoDBError.ItemError.DecodingError)
+//      .flatMap(FromAttributeValue.attrMapFromAttributeValue.fromAttributeValue)
 
 }
